Matrix-matrix Multiplication using mpi
The shared variable declared as follows: Double a[n,n], b[n,n], c[n,n]; The computation of the matrix multiplication follows the sequential program; The inner loop (with index k) computes the inner product of row I of matrix a and column j of matrix b, and then stores the result in c[i,j][1].Matrix multiplication is an example of embarrassingly parallel application, because there are a multitude of operations that can be executed in parallel. Two operations can be executed in parallel if they are independent. Since the write sets for pair of inner product are disjoint, we could compute all of them in parallel. Alternatively, compute row of result in parallel, column of results is parallel, or blocks of rows or columns in parallel.
The speedup and the efficiency analysis is performed by the formula given below:
Speedup (S) = (Serial Execution Time) / (Parallel execution Time)
Efficiency E(%) = Speedup (S) /No of processers (P)